# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Bidkan Clean Architecture.",
        "title": "Bidkan Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v1/users/login": {
            "post": {
                "description": "เข้าสู่ระบบด้วยชื่อผู้ใช้และรหัสผ่านเพื่อรับ JWT Token สำหรับการเข้าถึง API อื่นๆ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "เข้าสู่ระบบ",
                "parameters": [
                    {
                        "description": "ข้อมูลการเข้าสู่ระบบ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "เข้าสู่ระบบสำเร็จ",
                        "schema": {
                            "$ref": "#/definitions/login.Response"
                        }
                    },
                    "401": {
                        "description": "ชื่อผู้ใช้หรือรหัสผ่านไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v1/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงข้อมูลโปรไฟล์ของผู้คนที่กำลังเข้าสู่ระบบอยู่โดยใช้ JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "ข้อมูลโปรไฟล์ส่วนตัว",
                "responses": {
                    "200": {
                        "description": "ดึงข้อมูลสำเร็จ",
                        "schema": {
                            "$ref": "#/definitions/me.Response"
                        }
                    },
                    "401": {
                        "description": "ไม่ได้รับอนุญาต (Token ไม่ถูกต้องหรือหมดอายุ)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v1/users/verify/firebase": {
            "post": {
                "description": "ตรวจสอบ idToken ที่ได้รับจาก Firebase (ฝั่ง Frontend) เพื่อเปิดใช้งานบัญชีผู้ใช้อย่างเป็นทางการ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "ยืนยันตัวตนผ่าน Firebase และเปิดใช้งานบัญชี",
                "parameters": [
                    {
                        "description": "ข้อมูลการยืนยัน Firebase",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/verify_firebase.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "เปิดใช้งานบัญชีสำเร็จ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Token ไม่ถูกต้อง หรือข้อมูลผิดพลาด",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/wallet/topup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "สร้างรายการธุรกรรมเติมเงินและรับ QR Code สำหรับชำระเงินผ่าน PromptPay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "ขอเติมเงินเข้า Wallet (PromptPay)",
                "parameters": [
                    {
                        "description": "ระบุจำนวนเงินที่ต้องการเติม",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topup.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "สร้างรายการเติมเงินสำเร็จ",
                        "schema": {
                            "$ref": "#/definitions/topup.Response"
                        }
                    }
                }
            }
        },
        "/v1/wallet/transfer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "โอนเงินจาก Wallet ของเราไปยังผู้ใช้อื่นโดยระบุเบอร์โทรศัพท์ผู้รับ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "โอนเงินให้เพื่อน (P2P)",
                "parameters": [
                    {
                        "description": "ข้อมูลการโอนเงิน",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transfer.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "โอนเงินสำเร็จ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ข้อมูลไม่ถูกต้อง หรือยอดเงินไม่พอ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายในเซิร์ฟเวอร์",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "login.Request": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "example": "somchai_jaidee"
                }
            }
        },
        "login.Response": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "me.Response": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "topup.Request": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
        "topup.Response": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "qr_code_url": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "transfer.Request": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "receiver_phone": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "string"
                }
            }
        },
...
        "verify_firebase.Request": {
            "type": "object",
            "required": [
                "id_token"
            ],
            "properties": {
                "id_token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer ' followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}